tests: Split the aklite into two test suites
Let's split the heavy aklite test suite into two,
one dedicated to rollback and the second for
other use-cases.
